Changed to use SDL_QueueAudio
This seems to be vastly better.  The issue with the callback method, is that for some reason it was not calling the callback function at a regular interval.  This allows the emulator itself to dictate when audio gets passed.

It does however have one issue, where if the emulator is running too fast, the audio will be more and more 'lagged' out relative to the picture, since it just keeps adding more and more audio data.