NesSound - further refinements to resampling
Still needs some tweaking, but in a pretty good state now.  The emulator will now resample at a dynamic rate.

This does cause issues if the SDL window is manipulated in any way.  Might be a good idea to pause the emulation and clear the buffers if anything happens disrupts the SDL window.