Experimental resampling
* Implementation of dynamic resampling algorithm

Varies pitch way too much right now, but if there is a way to bring the generated samples per sample interval, perhaps we can get it to sound better.

One plus, is it does eliminate the gaps between SDL sample requests

* Changes to capframerate

This will allow thread to yield while waiting

* Changed to use SDL_QueueAudio

This seems to be vastly better.  The issue with the callback method, is that for some reason it was not calling the callback function at a regular interval.  This allows the emulator itself to dictate when audio gets passed.

It does however have one issue, where if the emulator is running too fast, the audio will be more and more 'lagged' out relative to the picture, since it just keeps adding more and more audio data.

* Add autoexec file

Executes console command at load

* Delete autoexec.cfg

Delete so this isn't tracked

* Update .gitignore

Add autoexec to ignored files

* Add CgtGenericFunctions

Provide generic average and slope calculations functions

* Update CgtGenericFunctions.h

Revise to add an average function that returns a float directly

* switchFullScreen changes

* NesSound - further refinements to resampling

Still needs some tweaking, but in a pretty good state now.  The emulator will now resample at a dynamic rate.

This does cause issues if the SDL window is manipulated in any way.  Might be a good idea to pause the emulation and clear the buffers if anything happens disrupts the SDL window.

* Changes to CgtLib library

CgtGenericFunctions - Add incrementAndWrap function